requirements.txt
streamlit==1.25.0               
openai==0.28.0                  
PyPDF2==3.0.1           
langchain==0.2.11          
langchain_community==0.0.1    
chromadb==0.4.3                
python-dotenv==1.0.0     
pillow==9.5.0 
